refactor: migrate ExceptionHandlingMiddleware to Frank
