+++ src/CampFitFurDogs.Api/Verticals/Authentication/Login/AuthLoginEndpoint.cs
@@ -1,7 +1,7 @@
 using System.Text.Json;
 using CampFitFurDogs.Application.Settings;
 using CampFitFurDogs.Domain.Errors;
-using Frank.Api;
+using Frank.Abstractions;
 using Frank.Settings;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
