+++ src/CampFitFurDogs.Api/Verticals/Dogs/GetDogProfileEndpoint.cs
@@ -1,6 +1,5 @@
 using CampFitFurDogs.Application.Abstractions.Dogs.GetDogProfile;
 using Frank.Abstractions;
-using Frank.Api;
 
 namespace CampFitFurDogs.Api.Verticals.Dogs;
