+++ src/Frank.Api/EndpointDiscovery.cs
@@ -1,5 +1,6 @@
 using System.Collections.Concurrent;
 using System.Reflection;
+using Frank.Abstractions;
 using Microsoft.AspNetCore.Routing;
 
 namespace Frank.Api;
