+++ src/Frank.Infrastructure/ExceptionHandling/ServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-using Frank.Api.ExceptionHandling;
+using Frank.Integration.ExceptionHandling;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Frank.Api;
@@ -8,7 +8,7 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddExceptionHandling(this IServiceCollection services)
     {
         // IExceptionHandler implementations are auto-registered via [AutoRegister] on the interface.
-        services.AddSingleton<ExceptionHandlingRegistry>();
+        services.AddSingleton<ExceptionHandlerRegistry>();
         return services;
     }
 }
