+++ src/Frank.Testing/Endpoints/CurrentUserIdEndpoint.cs
@@ -1,9 +1,8 @@
 using System.Security.Claims;
+using Frank.Abstractions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
-
-using Frank.Api;
 
 namespace Frank.Testing.Endpoints;
