+++ src/Frank.Testing/Endpoints/HealthCheckEndpoint.cs
@@ -1,8 +1,7 @@
+using Frank.Abstractions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
-
-using Frank.Api;
 
 namespace Frank.Testing.Endpoints;
