+++ src/Frank.Testing/Endpoints/ThrowEndpoint.cs
@@ -1,7 +1,6 @@
+using Frank.Abstractions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
-
-using Frank.Api;
 
 namespace Frank.Testing.Endpoints;
