+++ tests/Frank.Api.Tests/Fakes/FakeEndpoint.cs
@@ -1,3 +1,4 @@
+using Frank.Abstractions;
 using Microsoft.AspNetCore.Routing;
 
 namespace Frank.Api.Tests.Fakes;
