+++ tests/Frank.Api.Tests/Fakes/FakeEndpoint2.cs
@@ -1,3 +1,4 @@
+using Frank.Abstractions;
 using Microsoft.AspNetCore.Routing;
 
 namespace Frank.Api.Tests.Fakes;
