+++ tests/Frank.TestUtilities/Endpoints/HealthEndpoint.cs
@@ -1,4 +1,4 @@
-using Frank.Api;
+using Frank.Abstractions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
